requirements.txt
aiogram==2.24

pandas==1.5.2
SQLAlchemy==1.4.45
python-dotenv==0.21.0